requirements.txt
requests
baidu-aip
